// src/app/_angel/visitor/visitors/visitor-grid/models/visitor.model.ts
export interface Visitor {
    Id: number;
    Gorevli: number;
    Bilgi: string;
    Ad: string;
    Soyad: string;
    kart: string;
    Plaka: string;
    Arac: string;
    UserID: string;
    UserlistID: number;
    KimlikTipi: number;
    KimlikTipiAd: string;
    Kimlikno: string;
    Firma: string;
    ZiyaretNedeni: string;
    ZiyaretNedeniId: number;
    Sicilidadsoyad: string;
    SicilId1: number;
    Giris: string; // Date olarak parse edilebilinir
    Cikis: string | null;
    yetkistr: string;
    giztarih: string;
    isgtarih: string;
    islemno: string;
    ziyaretid: number;
    ziyaretsahibi: string | null;
    durum: string | null;
    songecisnoktasi: string | null;
    ZOKod1: string;
    ZOKod2: string;
    ZOKod3: string;
    ZOKod4: string;
    ZOKod5: string;
    ZOKod6: string;
    ZOKod7: string;
    ZOKod8: string;
    ZOKod9: string;
    ZOKod10: string;
    ZOKod11: string;
    ZOKod12: string;
    kaydeden: string;
    rowHeight: number;
}